requirements.txt
deep-translator==1.11.4
gradio==5.34.2
pythainlp==5.1.0b1
python-dotenv==1.1.0